Renaming some files to be more consistent with class name. Implemented Horizontal Scrolling of Thumbnails. Made Continuities sized to their number of itmes Incrementing to c++17 Adding an animation error panel to highlight errors in animation. Saving and restoring layout
