+++ src/cc_omniview_canvas.h
@@ -56,6 +56,7 @@ struct MarcherInfo {
 
 class CCOmniView_Canvas : public wxGLCanvas {
     using super = wxGLCanvas;
+
 public:
     CCOmniView_Canvas(AnimationView* view, wxWindow* parent,
         CalChartConfiguration& config,
