+++ src/confgr.h
@@ -202,9 +202,7 @@ private:                                           \
     DECLARE_CONFIGURATION_FUNCTIONS(CommandUndoSetSheet, bool);
     DECLARE_CONFIGURATION_FUNCTIONS(CommandUndoSelection, bool);
 
-    DECLARE_CONFIGURATION_FUNCTIONS(FieldFrameFieldThumbnailVisibility, bool);
-    DECLARE_CONFIGURATION_FUNCTIONS(FieldFrameFieldControlsVisibility, bool);
-    DECLARE_CONFIGURATION_FUNCTIONS(FieldFrameContinuityInfoVisibility, bool);
+    DECLARE_CONFIGURATION_FUNCTIONS(CalChartFrameAUILayout, wxString);
 
     DECLARE_CONFIGURATION_FUNCTIONS(ContCellLongForm, bool);
     DECLARE_CONFIGURATION_FUNCTIONS(ContCellFontSize, long);
