+++ src/contgram.y
@@ -30,6 +30,10 @@
 
 #include <vector>
 
+#if __cplusplus > 199711L
+#define register      // Deprecated in C++11.
+#endif  // #if __cplusplus > 199711L
+
 //#define YYDEBUG 1
 
 int yyerror(const char *s);
