+++ src/core/animatecompile.h
@@ -47,6 +47,9 @@ public:
     void RegisterError(AnimateError err, int line, int col, unsigned curr_pt,
         SYMBOL_TYPE contsymbol);
 
+    bool operator==(AnimationErrors const& rhs) const {
+        return mErrorMarkers == rhs.mErrorMarkers;
+    }
 private:
     std::map<AnimateError, ErrorMarker> mErrorMarkers;
 };
