+++ src/field_frame_controls.h
@@ -31,10 +31,10 @@ class FieldFrameControls : public wxPanel {
 public:
     // CalChartFrame will own the show that is passed in
     FieldFrameControls(double zoom, wxWindow* parent, wxWindowID id = wxID_ANY,
-    const wxPoint& pos = wxDefaultPosition,
-    const wxSize& size = wxDefaultSize,
-    long style = wxTAB_TRAVERSAL | wxNO_BORDER,
-    const wxString& name = wxPanelNameStr);
+        const wxPoint& pos = wxDefaultPosition,
+        const wxSize& size = wxDefaultSize,
+        long style = wxTAB_TRAVERSAL | wxNO_BORDER,
+        const wxString& name = wxPanelNameStr);
     virtual ~FieldFrameControls() = default;
 
     std::pair<CalChart::Coord::units, CalChart::Coord::units> GridChoice() const;
